// backend/src/generated/prisma/models/Message.ts
export type MessageModel = runtime.Types.Result.DefaultSelection<Prisma.$MessagePayload>
// ...
export type $MessagePayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Message"
  objects: {
    chat: Prisma.$ChatPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    chatId: string
    role: $Enums.MessageRole
    content: string
    status: $Enums.MessageStatus
    createdAt: Date
  }, ExtArgs["result"]["message"]>
  composites: {}
}